Support for element-based properties
Elements containing java.scene.Node descendants don't have to contain sub-nodes in the <children> tag.  By default, they will be added to the children collection.  Any node that contains children must be a descendant of Region.
Lots of rejigging around property management to cater for property setting in FXML elements as well as attributes.
Any class that is immutable is a special subclass of FXML element.  Some may yet be missing.